@@ -1,12 +1,74 @@
 <template>
   <div id="app">
-    <div id="nav">
+    <!-- <div id="nav">
       <router-link to="/">Home</router-link> |
       <router-link to="/about">About</router-link>
     </div>
-    <router-view/>
+    <router-view/> -->
+    <LuckyWheel
+    style="width: 800px; height: 800px"
+    ref="LuckyWheel"
+    :prizes="prizes"
+    :default-style="defaultStyle"
+    :blocks="blocks"
+    :buttons="buttons"
+    @start="startCallBack"
+    @end="endCallBack"
+  />
   </div>
 </template>
+
+<script>
+export default {
+  data () {
+    return {
+      prizes: [],
+      defaultStyle: {
+        fontColor: '#d64737',
+        fontSize: '14px'
+      },
+      blocks: [
+        { padding: '13px', background: '#d64737' }
+      ],
+      buttons: [
+        { radius: '50px', background: '#d64737' },
+        { radius: '45px', background: '#fff' },
+        { radius: '41px', background: '#f6c66f', pointer: true },
+        {
+          radius: '35px', background: '#ffdea0',
+          imgs: [{ src: require('./assets/logo.png'), width: '65%', top: '-50%' }]
+        }
+      ],
+    }
+  },
+  mounted () {
+    this.getPrizesList()
+  },
+  methods: {
+    getPrizesList () {
+      this.prizes = []
+      let data = [ '真皮长靴', '羊绒大衣', '保湿面霜','COACH蔻驰(二选一)', '皮草外套']
+      data.forEach((item, index) => {
+        this.prizes.push({
+          title: item,
+          background: index % 2 ? '#f9e3bb' : '#f8d384',
+          fonts: [{ text: item, top: '10%' }],
+          imgs:[{ src: require(`./assets/${index}.jpg`), width: '30%', top: '20%' }],
+        })
+      })
+    },
+    startCallBack () {
+      this.$refs.LuckyWheel.play()
+      setTimeout(() => {
+        this.$refs.LuckyWheel.stop(Math.random() * 5 >> 0)
+      }, 3000)
+    },
+    endCallBack (prize) {
+      alert(`恭喜你获得${prize.title}`)
+    },
+  }
+}
+</script>
 
 <style>
 #app {
